organization: Reorganize files
