+++ src/python/tests/test_todo.py
@@ -1,11 +1,8 @@
-import sys
-import os
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+from ..todo import TodoList
 
 import unittest
 from datetime import datetime
 from unittest.mock import Mock, patch
-from todo import TodoList
 
 class TestTodoList(unittest.TestCase):
     def setUp(self):
